requirements.txt
pandas==1.3.2
numpy==1.20.3
transformers==4.9.2
tensorboard==2.6.0
torch==1.11.0
sklearn==0.0
nltk==3.6.2
Flask==2.0.1
Flask-Cors==3.0.10
beautifulsoup4==4.9.3
pytorch-nlp==0.5.0
tqdm